requirements.txt
flask==3.0.0
openai==0.28.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pillow==10.1.0
beautifulsoup4==4.12.2
joblib==1.3.2
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.5.2 
flask-cors==4.0.0 
flask[async]==3.0.0 
gunicorn==21.2.0 
requests
